refactor: move EmploymentPreview component to employment-preview.tsx file
- Move the component to its own file
- Update the import path to the component in src/Preview.tsx
