+++ src/components/employment/employment.tsx
@@ -93,21 +93,3 @@ export function Employment() {
     </Details>
   );
 }
-
-export function EmploymentPreview() {
-  const position = useStore((state) => state.position);
-  const employer = useStore((state) => state.employer);
-  const startDate = useStore((state) => state.startDate);
-  const endDate = useStore((state) => state.endDate);
-  const description = useStore((state) => state.description);
-
-  return (
-    <>
-      {position}
-      {employer}
-      {startDate}
-      {endDate}
-      {description}
-    </>
-  );
-}
